requirements.txt
pygame
pillow
opencv-python
numpy
mediapipe
tensorflow